proper handling procedures

+++ packages/api/src/router/post.ts
@@ -1,9 +1,10 @@
-import { router, publicProcedure } from "../trpc";
+import { router, protectedProcedure } from "../trpc";
 import { z } from "zod";
-import { PostType } from "@prisma/client";
+import { PostType, Roles } from "@prisma/client";
+import { TRPCError } from "@trpc/server";
 
 export const postRouter = router({
-  create: publicProcedure
+  create: protectedProcedure
     .input(
       z.object({
         createdBy: z.string(),
@@ -20,13 +21,24 @@ export const postRouter = router({
     .mutation(({ ctx, input }) => {
       return ctx.prisma.post.create({ data: input });
     }),
-  all: publicProcedure.query(({ ctx }) => {
+  all: protectedProcedure.query(({ ctx }) => {
     return ctx.prisma.post.findMany();
   }),
-  byId: publicProcedure.input(z.string()).query(({ ctx, input }) => {
+  byId: protectedProcedure.input(z.string()).query(({ ctx, input }) => {
     return ctx.prisma.post.findFirst({ where: { id: input } });
   }),
-  deleteById: publicProcedure.input(z.string()).query(({ ctx, input }) => {
-    return ctx.prisma.post.delete({ where: { id: input } });
-  }),
+  deleteById: protectedProcedure
+    .input(z.string())
+    .query(async ({ input, ctx }) => {
+      const getPost = await ctx.prisma.post.findUnique({
+        where: { id: input },
+      });
+      if (!getPost) throw new TRPCError({ code: "NOT_FOUND" });
+      if (
+        getPost.createdBy !== ctx.session.user.id &&
+        ctx.session.user.role !== Roles.ADMIN
+      )
+        throw new TRPCError({ code: "FORBIDDEN" });
+      return ctx.prisma.post.delete({ where: { id: input } });
+    }),
 });
